MCR-2363 moved current servletjob code to MCRFrontendUtil
