+++ mycore-base/src/main/java/org/mycore/frontend/servlets/MCRContainerLoginServlet.java
@@ -96,7 +96,7 @@ public class MCRContainerLoginServlet extends MCRServlet {
 
         protected Optional<HttpServletRequest> getCurrentRequest() {
             LogManager.getLogger(getClass()).debug("Getting request from session: {}", session.getID());
-            return MCRServletJob.getCurrent()
+            return MCRFrontendUtil.getCurrentServletJob()
                 .map(MCRServletJob::getRequest);
         }
     }
